# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Racoon Bank Core Api",
    "version": "v1"
  },
  "paths": {
    "/api/bank-accounts/my": {
      "get": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Gets all bank accounts of a user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bank-accounts/{id}/history": {
      "get": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Gets operation history for a bank account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountOperationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountOperationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountOperationDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{id}/bank-accounts": {
      "get": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Gets all bank accounts of a user (for employee)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bank-accounts/all": {
      "get": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Gets all bank accounts (for employee)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccountDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bank-accounts": {
      "post": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Creates new bank account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBankAccountDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBankAccountDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBankAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bank-accounts/{id}": {
      "delete": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Deletes a bank account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/bank-accounts/{id}/deposit": {
      "put": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Deposits money to a bank account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Idempotency key for safe retries",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bank-accounts/{id}/withdraw": {
      "put": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Withdraws money from a bank account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Idempotency key for safe retries",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bank-accounts/transfer": {
      "put": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Transfer money from one bank account to another one",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Idempotency key for safe retries",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TransferDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              }
            }
          }
        }
      }
    },
    "/internal/bank-accounts/{id}/apply-credit": {
      "put": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Apply credit to a bank account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              }
            }
          }
        }
      }
    },
    "/internal/bank-accounts/{id}/pay-credit": {
      "put": {
        "tags": [
          "BankAccount"
        ],
        "summary": "Pay credit from a bank account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyOperationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BankAccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "accountNumber": {
            "type": "string",
            "nullable": true
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "BankAccountOperationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "$ref": "#/components/schemas/OperationType"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CreateBankAccountDto": {
        "required": [
          "currency"
        ],
        "type": "object",
        "properties": {
          "currency": {
            "$ref": "#/components/schemas/Currency"
          }
        },
        "additionalProperties": false
      },
      "Currency": {
        "enum": [
          "RUB",
          "USD",
          "EUR"
        ],
        "type": "string"
      },
      "MoneyOperationDto": {
        "type": "object",
        "properties": {
          "amount": {
            "minimum": 0.01,
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "OperationType": {
        "enum": [
          "Deposit",
          "Withdraw",
          "CreditIssued",
          "CreditPayment",
          "TransferOut",
          "TransferIn"
        ],
        "type": "string"
      },
      "TransferDto": {
        "type": "object",
        "properties": {
          "fromAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "toAccountNumber": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "amount": {
            "minimum": 0.01,
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}